fix: cast doodson value to int8_t for clarity in xdo_numerical function

+++ include/fes/xdo.hpp
@@ -39,7 +39,7 @@ inline auto xdo_numerical(const Eigen::Ref<const Vector7b>& doodson)
   // First element uses raw value, others add offset of 5
   result.push_back(code(doodson[0]));
   for (int64_t ix = 1; ix < doodson.size(); ++ix) {
-    result.push_back(code(doodson[ix] + 5));
+    result.push_back(code(static_cast<int8_t>(doodson[ix] + 5)));
   }
   return result;
 }
